refactor(js): Moved slack files to a slack folder within the channel folder
